* change evalutaion format- inside the train session instead of at the end

+++ src/main/java/problem/utils/Logger.java
@@ -88,7 +88,7 @@ public class Logger {
         // this.m_learningCurveDisplay.setActiveSeries(seriesName);
     }
 
-    public void addSeriesTime(long timeInSecs) {
+    public void addSeriesTime(double timeInSecs) {
         this.info("run time: " + timeInSecs + " secs");
         //this.m_learningCurveDisplay.addSeriesTime(timeInSecs);
     }
